Version finale CreateBknews, UpdateBknews, DeleteBknews

+++ FlOpEDT/api_graphql/bknews/delete.py
@@ -1,4 +1,3 @@
-from graphene_django import DjangoObjectType
 import graphene
 from displayweb.models import BreakingNews
 from .types import BknewsType
@@ -6,13 +5,17 @@ from graphql_relay import from_global_id
 
 class DeleteBknews(graphene.Mutation):
         class Arguments:
-            id = graphene.ID()
+            id = graphene.ID(required=True)
 
         bknews = graphene.Field(BknewsType)
 
         @classmethod
         def mutate(cls, root, info, id):
-            _, object_id = from_global_id(id)
-            bknews = BreakingNews.objects.get(id=object_id)
-            bknews.delete()
-            return DeleteBknews(bknews)
+            id = from_global_id(id) [1]
+            bknews_set = BreakingNews.objects.filter(id=id)
+            if bknews_set:
+                bknews=bknews_set.first()
+                bknews.delete()
+                return DeleteBknews(bknews)
+            else:
+                print('Breaking new with given ID does not exist in the database')
